Unify the color functions in color_tools to the color_utils file
